+++ Eclipse/src/ECS/SystemManager/Systems/System/MaterialSystem/MaterialSystem.cpp
@@ -43,7 +43,7 @@ namespace Eclipse
             engine->MaterialManager.StencilBufferClear();
         }
 
-        engine->Timer.tracker.system_end = glfwGetTime();
+        engine->Timer.tracker.system_end = static_cast<float>(glfwGetTime());
         engine->Timer.UpdateTimeContainer(engine->Timer.tracker);
 
         FrameMark
